storeのテストを修正
