+++ frontend/jest.config.js
@@ -3,7 +3,8 @@ module.exports = {
   collectCoverage: true,
   collectCoverageFrom: [
     '<rootDir>/components/**/*.vue',
-    '<rootDir>/pages/**/*.vue'
+    '<rootDir>/pages/**/*.vue',
+    '<rootDir>/store/**/*.js'
   ],
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/$1',
