+++ frontend/store/index.js
@@ -1,8 +1,8 @@
 const cookieparser = process.server ? require("cookieparser") : undefined;
 
 export const actions = {
-  nuxtServerInit({ commit }, { req }) {
-    if (req.headers.cookie) {
+  nuxtServerInit ({ commit }, { req }) {
+    if (cookieparser !== undefined && req.headers.cookie) {
       const parsed = cookieparser.parse(req.headers.cookie);
       try {
         const auth_flag = parsed.uid ? true : false;
